Added File Saving Functions
I added a fonction to save lists to files, set the file locations. And i changed the list structure and add a feature called listnumber this is an integer and it is a uniqe number that assigned it while it is forming. And when a list is created it updates a text file with the new value so it can be stored and past onto new running sessions.