Replaced the Master List concept with Main List
I deleted the concept of Master List and added the Main List concept instead. Main List logs the data in to a txt file instead of making them a linked list. I added a function to add list addresses as logs to Main List. Changed some existing code to fit with the new concept. Wrote the names of functions I need to code on this concept. Ran some tests to see if they are working or not. Reorganized the file structure for easy access.